refractor TCMD config into a single file
